+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.36.15](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.36.14...@carbon/ibm-cloud-cognitive@0.36.15) (2021-05-05)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.36.14](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.36.13...@carbon/ibm-cloud-cognitive@0.36.14) (2021-05-05)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.36.13](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.36.12...@carbon/ibm-cloud-cognitive@0.36.13) (2021-05-05)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.36.12](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.36.11...@carbon/ibm-cloud-cognitive@0.36.12) (2021-05-05)
+
+
+### Bug Fixes
+
+* page header deprecations ([#697](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/697)) ([ee559f9](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/ee559f93896a416c01d0e790864835542726d44a))
+
+
+
+
+
+## [0.36.11](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.36.10...@carbon/ibm-cloud-cognitive@0.36.11) (2021-05-04)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
 ## [0.36.10](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.36.9...@carbon/ibm-cloud-cognitive@0.36.10) (2021-04-30)
 
 **Note:** Version bump only for package @carbon/ibm-cloud-cognitive